modify updateTest part

+++ src/main/java/api/personalitytest/service/test/TestService.java
@@ -7,14 +7,10 @@ import api.personalitytest.domain.test.TestRepository;
 import api.personalitytest.web.dto.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -35,7 +31,8 @@ public class TestService {
     }
 
     @Transactional
-    public String save(UserSaveRequestDto userSaveRequestDto, ArrayList<ResultSaveRequestDto> resultSaveRequestDtoList,
+    public String save(UserSaveRequestDto userSaveRequestDto,
+                       ArrayList<ResultSaveRequestDto> resultSaveRequestDtoList,
                        ArrayList<ItemSaveRequestDto> itemSaveRequestDtoList) {
 
         Test savedTest = testRepository.save(userSaveRequestDto.toEntity());
@@ -118,32 +115,33 @@ public class TestService {
     }
 
     @Transactional
-    public SuccessResponseDto update(Long testId,
+    public String update(Long testId,
                                      TestUpdateRequestDto testDto,
-                                     ArrayList<ResultUpdateRequestDto> resultUpdateDto,
-                                     ArrayList<ItemUpdateRequestDto> itemUpdateDto,
-                                     MultipartFile file,
-                                     String dir) throws IOException {
+                                     ArrayList<ResultUpdateRequestDto> resultDtoList,
+                                     ArrayList<ItemUpdateRequestDto> itemDtoList)
+    {
+        Optional<Test> optionalTest = testRepository.findById(testId);
 
-        Test test = testRepository.findById(testId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 테스트가 없습니다. + testId" + testId));
+        if (optionalTest.isEmpty()) {
+            return "";
+        }
+
+        Test test = optionalTest.get();
 
         // Dirty Checking
         test.update(testDto);
 
         // 기존 테스트 데이터 및 결과 우선 삭제 하고 새로 저장
         resultRepository.deleteResultsByTestId(testId);
-        for (ResultUpdateRequestDto dto : resultUpdateDto) {
+        for (ResultUpdateRequestDto dto : resultDtoList) {
             resultRepository.save(dto.toEntity(testId));
         }
 
         itemRepository.deleteItemsByTestId(testId);
-        for (ItemUpdateRequestDto dto : itemUpdateDto) {
+        for (ItemUpdateRequestDto dto : itemDtoList) {
             itemRepository.save(dto.toEntity(testId));
         }
 
-        file.transferTo(new File(dir + test.getFullImageName()));
-
-        return new SuccessResponseDto(true);
+        return test.getFullImageName();
     }
 }
